Minor API change: allow sending _no_ messages
The `Protobuf` interface for client and bidirectional streaming took a function

``` haskell
IO (IsFinal, Input (RPC s m))
```

which it repeatedly called until no more inputs needed to be sent. However,
this meant that we could not send _no_ inputs; the API is now

``` haskell
IO (IsFinal, Maybe (Input (RPC s m)))
```

The contract here is that `(Final, Nothing)` should _only_ be used in the case
there are no messages to be sent at all.

This matches the (already present) detailed discussion of the quirks of
`sendInput` in `Network.GRPC.Client`.