Allow to define some methods using `RawHandler`
This gives full access to the server API, which can for example be used for
metadata. `wait_for_ready_with_client_timeout_example_client.py` in the gRPC
repo now returns the exact same results with the example server as with the
`grapesy` example.

This is the first step towards the final (for now) step towards a smoother API:
removing metadata from the convenience handlers, offering it only through the
full API.