Introduce `DataFlow`
This is a minor abstraction, which captures the data flow over a session:
headers, then messages, then trailers. This allows us to talk about an "inbound
dataflow" and an "outbound dataflow". This by itself is merely a nice-to-have,
but it sets the stage for the next step, which is to properly deal with the
gRPC Trailers-Only case.