Continue test-suite
The dialogue-based test suite is _so_ useful for discovering inconsistencies
between the server and the client API, since a _single_ description of the
communication is interpreted in "dual" ways by both the server and the client.

We still haven't actually _executed_ any of these tests (we're getting close to
that point now), but simply writing them has already proven very useful indeed.