Potential solution
Turns out, the HTTP2 exception at its root _is_ a blocked indefinitely
exception, which means we are in fact in the two-thread-waiting-for-each-other
case, for which we have a workaround. Still need to figure out if the
workaround works for the full thing also.

(It would also be better if we patched HTTP2 so that the underlying exception
is _not_ a "blocked indefinitely" exception.)