Be more precise about request handler exceptions
The key type introduced in this commit is

```haskell
-- | HTTP2 request handler
--
-- The request handler should not throw any exceptions. Most of the time we will
-- therefore work with
--
-- > RequestHandler (Either SomeException ())
--
-- It is passed a function to unmask asynchronous exceptions when it is ready.
type RequestHandler a =
     (forall x. IO x -> IO x)
  -> HTTP2.Request
  -> (HTTP2.Response -> IO ())
  -> IO a
```

`ServerParams` now has a new field

```haskell
serverTopLevel ::
     RequestHandler (Either SomeException ())
  -> RequestHandler ()
```

This replaces

```haskell
serverHandlerHook :: HTTP2.Server -> HTTP2.Server
```

in `ServerSetup`; in fact, `ServerSetup` is now gone entirely (along with the
corresponding `ServerSetup` field in `ServerConfig`).

To hide some of these changes, a new function `runServerWithHandlers` combines
`mkGrpcServer` and `runServer`; client code is encouraged to use this function.

With these changes in place, we no longer need the partial duplication between
the server exception logger and the handler lock in the test suite; in fact,
the exception logger is now removed entirely. Previously, the handler lock was
using `serverHandlerHook` but the type of `serverHandlerHook` was misleading,
as it wasn't actually able to catch exceptions -- they were swallowed earlier
in the pipeline. The new setup is much more explicit about which exceptions can
happen where.