No more custom expected expections
This is the first step towards making exception handling in the test suite more
surgical: rather than collecting exceptions at the top-level, we verify _at the
source_ whether or not they are expected, and if so, don't allow them to bubble
up at all. This means that we can then stop the test suite at the first
(unexpected) exception. This will make the tests easier to understand and
setup: for example, as things stand, when one client fails, the other clients
are killed, but that will then result in further exceptions in those clients.

This doesn't get us all the way there yet, but it makes some progress towards
this goal.