Fix cancellation interop test
This was pretty subtle: we were waiting for the inbound thread to die, but it
never got out of initializing state because the _outbound_ thread was killed
too soon, before it could give `http2` the nudge it requires to start the
request. Fixing this required changing the semantics of the `Thread`
abstraction; see discussion in `ThreadNotStarted`.