Split off `Supports{Client/Server}Rpc`
This can be useful for example when wanting to write a client for an RPC with
an existing server; by splitting off these two classes the methods that are
only required by the server don't need to be defined. This isn't too useful
_yet_, because for messages we typically define the format once and for all and
most client code doesn't need to worry about it. However, this paves the way
for the introduction of typed metadata, which is application specific and so
will need application specific instances.