Provide access to full request headers in server
This means that when we add additional fields (such as `requestTraceContext` or
`requestPreviousRpcAtttempts`), we do not necessarily need to add specific API
for accessing these fields (until we properly implement support for
OpenTelemetry or automatic retries, respectively).

On the client side, the same is offered through `recvInitialResponse` as well
as `recvOutputWithEnvelope`, which provides access to the raw response headers
and trailers, thereby giving access to specialized fields such as
`properTrailersPushback` (for retries) and `properTrailersOrcaLoadReport` (for
load balancing).