Disable HTTP-level timeouts
We can still see `TimeoutThread` exceptions if a server is terminated before
our worker gets a chance to install its own exception handler. This is not a
big deal, but can lead to some confusing error messages for example in the
`interop` tests: if a test fails, and the server is terminated very quickly,
then this exception may be shown (we could hide it by calling
`setUncaughtExceptionHandler`).