Initial stress tests, fix `ThreadDone` leak
Added stress tests for many non-streaming calls on one connection and many
connections with one non-streaming call. Left a stub to fill in for "many
messages" stress test.

In the many calls case, we discovered a memory leak caused by retention of
`ThreadDone` objects in `startRPC`, which this commit also fixes.

See #133 for a (less critical) memory issue we discovered in the many connection
case.