Explicitly use `127.0.0.1` in test suite
When `Nothing` is used as a host, it gets defaulted to `localhost`, which
resolves differently in IPv6 vs IPv4.

Remove TlsFailHostName test, because it depends on having two names for the same
server (`localhost` vs `127.0.0.1`), but the choice of two names depends on what
`localhost` resolves to. The test doesn't add much value, since we still verify
that the certificate is rejected if it is not set as a root certificate.