Remove `XIO`
It served its purpose, allowing us to be more precise about which exceptions we
want to handle here. Although it has not yet unlived its usefulness, it was
also poorly motivated, and there were concerns about unsoundness (what happens
if you try to `liftIO takeMVar`?). We could fix it, and perhaps one day we
should, but for now we just remove it altogether; it's not a problem we must
solve for this project.