Wrap outbound client thread stream exceptions
Unlike the server inbound/outbound and client outbound threads, `http2` is aware
of the client outbound thread. If the server disconnects, there is a race
between the `http2`-thrown exception and the exception that will come from
`grapesy` attempting to read. No matter who wins that race, we want to mark the
exception with `ServerDisconnected`.