Fix `Test.Sanity.Disconnect`
The test was flaky: if the server was killed too early, we might not get to see
the number of disconnects we were expecting. This significantly refactors the
tests to be much more precise; I cannot reproduce any test failures now.

This also fixes an unrelated bug in `echoHandler`, which was installing a new
exception handler on each loop iteration.

Closes #248.