Remove `HTTP2Settings` from `ServerParams`
This was a relatively common pattern:

```haskell
server <- mkGrpcServer serverParams $ ..
forkServer serverParams serverConfig server $ \runningServer -> ..
```

but it is misleading: notice that the `serverParams` need to be passed _twice_,
even though on call to `forkServer` we need _only_ the `HTTP2Settings`, and on
the call to `mkGrpcServer` we need everything in `ServerParams` _except_ the
`HTTP2Settings`: they are completely separate. This gets even more confusing in
a pattern such as this:

```haskell
server <- mkGrpcServer def $ ..
forkServer def serverConfig server $ \runningServer -> ..
```

where we are using `def` in both places; becomes quite easy to make a change in
the wrong place, and then be confused that it didn't take effect. We are now
separating these two values out completely.